pi_hole_web_survillance/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "pihole-analytics"
version = "1.0.0"
description = "Network surveillance and analytics suite for Pi-hole"
requires-python = ">=3.9"
dependencies = [
    "flask>=3.0",
    "flask-limiter>=3.5",
    "pyyaml>=6.0",
    "requests>=2.31",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
pihole-fetch     = "scripts.data.fetcher:run_fetch"
pihole-report    = "scripts.core.reporter:main"
pihole-dashboard = "scripts.web.app:main"
pihole-download  = "scripts.data.downloader:run_download"

[tool.setuptools.packages.find]
where = ["."]
include = ["scripts*"]
